# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "PDPS Backend API",
        "description": "API documentation for PDPS (Pradeshiya Sabha Development Project System) Backend",
        "contact": {
            "email": "admin@pdps.lk"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Development Server"
        }
    ],
    "paths": {
...
        "/news": {
            "get": {
                "tags": [
                    "News"
                ],
                "summary": "Get all news",
                "description": "Retrieve a list of all news articles",
                "operationId": "21b84dbdd75fd20b9c9ddb9b9e68f0a1",
                "responses": {
                    "200": {
                        "description": "News retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "AllNews": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "news_en": {
                                                        "type": "string",
                                                        "example": "English news content"
                                                    },
                                                    "news_si": {
                                                        "type": "string",
                                                        "example": "Sinhala news content"
                                                    },
                                                    "news_ta": {
                                                        "type": "string",
                                                        "example": "Tamil news content"
                                                    },
                                                    "priority": {
                                                        "type": "integer",
                                                        "example": 1
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "News"
                ],
                "summary": "Create a new news article",
                "description": "Add a new news article to the system",
                "operationId": "d6a12c2e3b4ea9ab588ba9b1cb14748f",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "newsSinhala",
                                    "newsEnglish",
                                    "newsTamil"
                                ],
                                "properties": {
                                    "newsSinhala": {
                                        "type": "string",
                                        "example": "සිංහල පුවත් අන්තර්ගතය"
                                    },
                                    "newsEnglish": {
                                        "type": "string",
                                        "example": "English news content"
                                    },
                                    "newsTamil": {
                                        "type": "string",
                                        "example": "தமிழ் செய்தி உள்ளடக்கம்"
                                    },
                                    "priority": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "News article created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "News article created successfully"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/news/{id}": {
            "put": {
                "tags": [
                    "News"
                ],
                "summary": "Update a news article",
                "description": "Update an existing news article",
                "operationId": "25ea273290c6cb379ab85e08ff4fadfc",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "News article ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "newsSinhala",
                                    "newsEnglish",
                                    "newsTamil"
                                ],
                                "properties": {
                                    "newsSinhala": {
                                        "type": "string",
                                        "example": "Updated Sinhala news content"
                                    },
                                    "newsEnglish": {
                                        "type": "string",
                                        "example": "Updated English news content"
                                    },
                                    "newsTamil": {
                                        "type": "string",
                                        "example": "Updated Tamil news content"
                                    },
                                    "priority": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "News article updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "News article updated successfully"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "News article not found"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "News"
                ],
                "summary": "Delete a news article",
                "description": "Delete an existing news article",
                "operationId": "bbe79a918bfbb58bc46ed74b43e9b986",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "News article ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "News article deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "News deleted successfully."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "News article not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "News not found."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/newsCount": {
            "get": {
                "tags": [
                    "News"
                ],
                "summary": "Get news count",
                "description": "Get the total count of news articles",
                "operationId": "e004b4fd10134748181401268a358bac",
                "responses": {
                    "200": {
                        "description": "News count retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "count": {
                                            "type": "integer",
                                            "example": 25
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/siteNewsView": {
            "get": {
                "tags": [
                    "News"
                ],
                "summary": "Get news for site view",
                "description": "Get news articles for public site view in specified language",
                "operationId": "1b6b3ebb1165d3dd28ec27c15a76e0a6",
                "parameters": [
                    {
                        "name": "language",
                        "in": "query",
                        "description": "Language code",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "en",
                                "si",
                                "ta"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "News retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "500": {
                        "description": "Server error"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Enter your API token in the format: Bearer {token}",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
    "security": [
        {
            "sanctum": []
        }
    ]
}